separate usage semantics
